content-security-policy/sandbox/service-worker-sandbox.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<body>
<script>
let frame = null;
let worker = null;
const scope = 'support/empty.html';
const script = 'support/sandboxed-service-worker.js';

// Global setup: this must be the first promise_test.
promise_test(async (t) => {
  const registration =
      await service_worker_unregister_and_register(t, script, scope);
  worker = registration.installing;
  await wait_for_state(t, worker, 'activated');
  frame = await with_iframe(scope);
}, 'global setup');

promise_test(async (t) => {
  const r = await frame.contentWindow.fetch('/get-origin', {mode: 'cors'});
  const j = await r.json();
  assert_equals(j.origin, 'null', 'Origin should be opaque');
}, 'Origin of sandboxed service worker');

promise_test(async (t) => {
  const r = await frame.contentWindow.fetch('/get-origin', {mode: 'same-origin'});
  const j = await r.json();
  assert_equals(j.origin, 'null', 'Origin should be opaque');
}, 'Response generated by sandboxed service worker can be fetched as same-origin');

promise_test(t => {
  return promise_rejects_js(
      t,
      frame.contentWindow.TypeError,
      frame.contentWindow.fetch(
        '/fetch?url=' + encodeURIComponent(location.origin + '/fetch/api/resources/top.txt?hash=' + Math.random()),
        {mode: 'same-origin'}));
}, 'Fetch by sandboxed service worker should fail because of opaque origin (mode: same-origin)');

promise_test(t => {
  return promise_rejects_js(
      t,
      frame.contentWindow.TypeError,
      frame.contentWindow.fetch(
        '/fetch?url=' + encodeURIComponent(location.origin + '/fetch/api/resources/top.txt?hash=' + Math.random()),
        {mode: 'cors'}));
}, 'Fetch by sandboxed service worker should fail because of opaque origin (mode: cors)');

promise_test(t => {
  return promise_rejects_js(
      t,
      frame.contentWindow.TypeError,
      frame.contentWindow.fetch(
        '/fetch?url=' + encodeURIComponent(location.origin + '/fetch/api/resources/cors-top.txt?hash=' + Math.random()),
        {mode: 'same-origin'}));
}, 'Fetch by sandboxed service worker should fail because of opaque origin (mode: same-origin, with ACAOrigin)');

promise_test(async (t) => {
  const r = await frame.contentWindow.fetch(
    '/fetch?url=' + encodeURIComponent(location.origin + '/fetch/api/resources/cors-top.txt?hash=' + Math.random()),
    {mode: 'cors'});
  const text = await r.text();
  assert_equals(text, 'top');
}, 'Fetch by sandboxed service worker should succeed (mode: cors, with ACAOrigin)');

// Global cleanup: the final promise_test.
promise_test(async (t) => {
  if (frame)
    frame.remove();
  await service_worker_unregister(t, scope);
}, 'global cleanup');
</script>
